feat(services): AddressServiceImpl and ProductServiceImpl validations with Authentication

+++ src/main/java/com/school/ecommerceupch/services/AddressServiceImpl.java
@@ -15,6 +15,8 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
+
 @Service
 public class AddressServiceImpl implements IAddressService {
 
@@ -42,6 +44,12 @@ public class AddressServiceImpl implements IAddressService {
 
     @Override
     public BaseResponse get(Long id) {
+
+        UserDetailsImpl userAuthenticated = getUserAuthenticated();
+
+        if (!userAuthenticated.getUser().getId().equals(id))
+            throw new AccessDeniedException();
+
         Address address = findOneAndEnsureExistById(id);
 
         return BaseResponse.builder()
@@ -55,6 +63,9 @@ public class AddressServiceImpl implements IAddressService {
     @Override
     public BaseResponse update(Long id, UpdateAddressRequest request) {
         UserDetailsImpl userAuthenticated = getUserAuthenticated();
+
+        if (!userAuthenticated.getUser().getId().equals(id))
+            throw new AccessDeniedException();
 
         Address address = findOneAndEnsureExistById(id);
 
@@ -74,14 +85,18 @@ public class AddressServiceImpl implements IAddressService {
 
     @Override
     public BaseResponse delete(Long id) {
+        UserDetailsImpl userAuthenticated = getUserAuthenticated();
 
         if(!repository.existsById(id))
             throw new ObjectNotFoundException("Address not found");
 
+        if (!userAuthenticated.getUser().getId().equals(id))
+            throw new AccessDeniedException();
+
         repository.deleteById(id);
 
         return BaseResponse.builder()
-                .data(null)
+                .data(Collections.EMPTY_LIST)
                 .message("Address deleted correctly ")
                 .success(Boolean.TRUE)
                 .httpStatus(HttpStatus.NO_CONTENT)
@@ -103,6 +118,7 @@ public class AddressServiceImpl implements IAddressService {
         Address address = new Address();
         address.setStreet(request.getStreet());
         address.setZipcode(request.getZipcode());
+        address.setState(request.getState());
         address.setCountry(request.getCountry());
         address.setUser(user);
 
@@ -112,6 +128,7 @@ public class AddressServiceImpl implements IAddressService {
     private Address update(Address address, UpdateAddressRequest request) {
         address.setStreet(request.getStreet());
         address.setZipcode(request.getZipcode());
+        address.setState(request.getState());
         address.setCountry(request.getCountry());
         repository.save(address);
         return address;
